NCD-672: Move and rename DeviceSelector.tsx
